+++ lib/src/fql/expr.dart
@@ -8,7 +8,7 @@ part 'expr.g.dart';
 
 @JsonSerializable()
 class Expr {
-  static Object wrap_value(dynamic value) {
+  static Object? wrap_value(dynamic value) {
     if (value is List) {
       return wrap_values(value);
     } else if (value is Map<String, dynamic>) {
@@ -20,7 +20,7 @@ class Expr {
     }
   }
 
-  static Object wrap_values(Object data) {
+  static Object? wrap_values(Object? data) {
     if (data == null) return null;
 
     if (data is List) {