+++ lib/src/fql/read_and_write.g.dart
@@ -9,10 +9,8 @@ part of 'read_and_write.dart';
 Get _$GetFromJson(Map<String, dynamic> json) {
   $checkKeys(json, disallowNullValues: const ['ts']);
   return Get(
-    json['get'] == null
-        ? null
-        : Expr.fromJson(json['get'] as Map<String, dynamic>),
-    ts: json['ts'] == null ? null : DateTime.parse(json['ts'] as String),
+    json['get'] ?? Expr.fromJson(json['get'] as Map<String, dynamic>),
+    ts: json['ts'] ?? DateTime.parse(json['ts'] as String),
   );
 }
 
@@ -46,10 +44,8 @@ Paginate _$PaginateFromJson(Map<String, dynamic> json) {
   $checkKeys(json,
       disallowNullValues: const ['ts', 'after', 'size', 'events', 'sources']);
   return Paginate(
-    json['paginate'] == null
-        ? null
-        : Expr.fromJson(json['paginate'] as Map<String, dynamic>),
-    ts: json['ts'] == null ? null : DateTime.parse(json['ts'] as String),
+    json['paginate'] ?? Expr.fromJson(json['paginate'] as Map<String, dynamic>),
+    ts: json['ts'] ?? DateTime.parse(json['ts'] as String),
     size: json['size'] as int,
     before: json['before'],
     after: json['after'],
@@ -82,12 +78,9 @@ Select _$SelectFromJson(Map<String, dynamic> json) {
   $checkKeys(json, disallowNullValues: const ['default']);
   return Select(
     json['select'],
-    json['from'] == null
-        ? null
-        : Expr.fromJson(json['from'] as Map<String, dynamic>),
-    default_: json['default'] == null
-        ? null
-        : Expr.fromJson(json['default'] as Map<String, dynamic>),
+    json['from'] ?? Expr.fromJson(json['from'] as Map<String, dynamic>),
+    default_: json['default'] ??
+        Expr.fromJson(json['default'] as Map<String, dynamic>),
   );
 }
 
@@ -109,12 +102,8 @@ Map<String, dynamic> _$SelectToJson(Select instance) {
 
 Create _$CreateFromJson(Map<String, dynamic> json) {
   return Create(
-    json['create'] == null
-        ? null
-        : Expr.fromJson(json['create'] as Map<String, dynamic>),
-    json['params'] == null
-        ? null
-        : Obj.fromJson(json['params'] as Map<String, dynamic>),
+    json['create'] ?? Expr.fromJson(json['create'] as Map<String, dynamic>),
+    json['params'] ?? Obj.fromJson(json['params'] as Map<String, dynamic>),
   );
 }
 
@@ -125,9 +114,8 @@ Map<String, dynamic> _$CreateToJson(Create instance) => <String, dynamic>{
 
 CreateCollection _$CreateCollectionFromJson(Map<String, dynamic> json) {
   return CreateCollection(
-    json['create_collection'] == null
-        ? null
-        : Obj.fromJson(json['create_collection'] as Map<String, dynamic>),
+    json['create_collection'] ??
+        Obj.fromJson(json['create_collection'] as Map<String, dynamic>),
   );
 }
 
@@ -138,9 +126,8 @@ Map<String, dynamic> _$CreateCollectionToJson(CreateCollection instance) =>
 
 CreateDatabase _$CreateDatabaseFromJson(Map<String, dynamic> json) {
   return CreateDatabase(
-    json['create_database'] == null
-        ? null
-        : Obj.fromJson(json['create_database'] as Map<String, dynamic>),
+    json['create_database'] ??
+        Obj.fromJson(json['create_database'] as Map<String, dynamic>),
   );
 }
 
@@ -151,9 +138,8 @@ Map<String, dynamic> _$CreateDatabaseToJson(CreateDatabase instance) =>
 
 CreateFunction _$CreateFunctionFromJson(Map<String, dynamic> json) {
   return CreateFunction(
-    json['create_function'] == null
-        ? null
-        : Obj.fromJson(json['create_function'] as Map<String, dynamic>),
+    json['create_function'] ??
+        Obj.fromJson(json['create_function'] as Map<String, dynamic>),
   );
 }
 
@@ -164,9 +150,8 @@ Map<String, dynamic> _$CreateFunctionToJson(CreateFunction instance) =>
 
 CreateIndex _$CreateIndexFromJson(Map<String, dynamic> json) {
   return CreateIndex(
-    json['create_index'] == null
-        ? null
-        : Obj.fromJson(json['create_index'] as Map<String, dynamic>),
+    json['create_index'] ??
+        Obj.fromJson(json['create_index'] as Map<String, dynamic>),
   );
 }
 
@@ -177,9 +162,8 @@ Map<String, dynamic> _$CreateIndexToJson(CreateIndex instance) =>
 
 CreateKey _$CreateKeyFromJson(Map<String, dynamic> json) {
   return CreateKey(
-    json['create_key'] == null
-        ? null
-        : Obj.fromJson(json['create_key'] as Map<String, dynamic>),
+    json['create_key'] ??
+        Obj.fromJson(json['create_key'] as Map<String, dynamic>),
   );
 }
 
@@ -189,9 +173,8 @@ Map<String, dynamic> _$CreateKeyToJson(CreateKey instance) => <String, dynamic>{
 
 CreateRole _$CreateRoleFromJson(Map<String, dynamic> json) {
   return CreateRole(
-    json['create_role'] == null
-        ? null
-        : Obj.fromJson(json['create_role'] as Map<String, dynamic>),
+    json['create_role'] ??
+        Obj.fromJson(json['create_role'] as Map<String, dynamic>),
   );
 }
 
@@ -202,9 +185,7 @@ Map<String, dynamic> _$CreateRoleToJson(CreateRole instance) =>
 
 Delete _$DeleteFromJson(Map<String, dynamic> json) {
   return Delete(
-    json['delete'] == null
-        ? null
-        : Expr.fromJson(json['delete'] as Map<String, dynamic>),
+    json['delete'] ?? Expr.fromJson(json['delete'] as Map<String, dynamic>),
   );
 }
 
@@ -224,14 +205,10 @@ Map<String, dynamic> _$EventsToJson(Events instance) => <String, dynamic>{
 
 Insert _$InsertFromJson(Map<String, dynamic> json) {
   return Insert(
-    json['insert'] == null
-        ? null
-        : Ref.fromJson(json['insert'] as Map<String, dynamic>),
+    json['insert'] ?? Ref.fromJson(json['insert'] as Map<String, dynamic>),
     json['ts'],
     json['action'],
-    json['params'] == null
-        ? null
-        : Obj.fromJson(json['params'] as Map<String, dynamic>),
+    json['params'] ?? Obj.fromJson(json['params'] as Map<String, dynamic>),
   );
 }
 
@@ -269,9 +246,7 @@ Map<String, dynamic> _$MergeToJson(Merge instance) {
 
 Remove _$RemoveFromJson(Map<String, dynamic> json) {
   return Remove(
-    json['remove'] == null
-        ? null
-        : Expr.fromJson(json['remove'] as Map<String, dynamic>),
+    json['remove'] ?? Expr.fromJson(json['remove'] as Map<String, dynamic>),
     json['ts'],
     json['action'],
   );
@@ -285,12 +260,8 @@ Map<String, dynamic> _$RemoveToJson(Remove instance) => <String, dynamic>{
 
 Replace _$ReplaceFromJson(Map<String, dynamic> json) {
   return Replace(
-    json['replace'] == null
-        ? null
-        : Expr.fromJson(json['replace'] as Map<String, dynamic>),
-    json['params'] == null
-        ? null
-        : Obj.fromJson(json['params'] as Map<String, dynamic>),
+    json['replace'] ?? Expr.fromJson(json['replace'] as Map<String, dynamic>),
+    json['params'] ?? Obj.fromJson(json['params'] as Map<String, dynamic>),
   );
 }
 
@@ -301,12 +272,8 @@ Map<String, dynamic> _$ReplaceToJson(Replace instance) => <String, dynamic>{
 
 Update _$UpdateFromJson(Map<String, dynamic> json) {
   return Update(
-    json['update'] == null
-        ? null
-        : Expr.fromJson(json['update'] as Map<String, dynamic>),
-    json['params'] == null
-        ? null
-        : Obj.fromJson(json['params'] as Map<String, dynamic>),
+    json['update'] ?? Expr.fromJson(json['update'] as Map<String, dynamic>),
+    json['params'] ?? Obj.fromJson(json['params'] as Map<String, dynamic>),
   );
 }