+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.11.0] - 2021-03-13
+
+### Changed
+
+- Adjustments to null safety
+- Function `merge` has been deprecated due to null-safety changes.
+- Updated dependencies:
+    - http: ^0.13.0
+    - json_annotation: ^4.0.0
+- Updated development dependencies:
+    - json_serializable: ^4.0.2
+    - build_runner: ^1.12.1
+    - pedantic: ^1.11.0
+- Added dependency override to prevent version discrepancies:
+    - build: ^2.0.0
+
+
 ## [0.10.0] - 2021-03-10
 
 ### Added